% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_months.r
\name{add_months}
\alias{add_months}
\title{This is a date math function for months}
\usage{
add_months(date, add_months = 1)
}
\description{
This is a date math function for months
}
\examples{
add_months('2017-11-19')
add_months(your_boos_bday)
add_months()
}
\keyword{date,}
\keyword{month,}
\keyword{time}
