% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_names.r
\name{analysis_names}
\alias{analysis_names}
\title{This is a function to format variable names to make analysis easier}
\usage{
analysis_names(df)
}
\description{
This is a function to format variable names to make analysis easier
}
\examples{
analysis_names(your_df)
}
\keyword{column,}
\keyword{data}
\keyword{frame}
\keyword{names,}
\keyword{variable,}
