% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_factor.r
\name{count_factor}
\alias{count_factor}
\title{This function counts factors in a data frame for each variable}
\usage{
count_factor(df)
}
\description{
This function counts factors in a data frame for each variable
}
\examples{
count_factor(your_df)
}
\keyword{data}
\keyword{factor,}
\keyword{frame}
