% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_factors.r
\name{count_factors}
\alias{count_factors}
\title{This function counts factors in a data frame for each variable}
\usage{
count_factors(df)
}
\description{
This function counts factors in a data frame for each variable
}
\examples{
count_factors(your_df)
}
\keyword{data}
\keyword{factor,}
\keyword{frame}
