% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_na.r
\name{count_na}
\alias{count_na}
\title{This function counts NAs in a data frame for each variable}
\usage{
count_na(df)
}
\description{
This function counts NAs in a data frame for each variable
}
\examples{
count_na(your_df)
}
\keyword{data}
\keyword{frame}
\keyword{na,}
