% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummify.r
\name{dummify}
\alias{dummify}
\title{This is a function remove space characters from the head and tail of a string}
\usage{
dummify(data, id_var)
}
\description{
This is a function remove space characters from the head and tail of a string
}
\examples{
dummify(df, id_var = 'id')
}
\keyword{dummies,}
\keyword{dummy}
\keyword{dummy,}
\keyword{variable}
