% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{normalize}
\alias{normalize}
\title{This function normalize a values in a numeric variable}
\usage{
normalize(j)
}
\description{
This function normalize a values in a numeric variable
}
\examples{
normalize(df$numeric_variable)
normalize(string_of_numbers)
normalize(c(1,3,4))
}
\keyword{normalize,}
\keyword{numeric}
\keyword{one,}
\keyword{standardize,}
\keyword{zero,}
