% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_to_date.r
\name{num_to_date}
\alias{num_to_date}
\title{This function converts a six-digit numeric date (yyyymm) to a value in a proper date format}
\usage{
num_to_date(six_digit_date, day_of_month = 15)
}
\description{
This function converts a six-digit numeric date (yyyymm) to a value in a proper date format
}
\examples{
num_to_date(200105)
}
\keyword{date,}
\keyword{time}
