\name{showComposition}
\alias{showComposition}
\title{Show the chemical and physical combinations in the database}
\description{
Show the values of chemical composition (Z, Y, AFE) and mixing-length
present in the database.
 }
 \usage{
showComposition()
}
\value{
  Print the combinations of the inputs that exist in the database.
}
\seealso{
\code{\link{testComposition}}.
}

\references{
  
M. Dell'Omodarme, G. Valle, S. Degl'Innocenti, and P.G. Prada Moroni
(2012). \emph{The Pisa Stellar Evolution Data Base for low-mass
  stars}. Astronomy and Astrophysics, 540, A26.
}
\examples{

  showComposition()
}
\keyword{manip}
