% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{stars}
\alias{stars}
\title{stars}
\usage{
stars(text = "", user = NULL)
}
\arguments{
\item{text}{Text string to search for}

\item{user}{Name of github user whose stars you want to search. Defaults to
your own profile}
}
\value{
Interactive screen dump of results enabling you to select the best
match and open the corresponding github repository
}
\description{
Text search github stars
}
