\name{stellaR-package}
\docType{package}
\alias{stellaR-package}
\title{Stellar tracks and isochrones from Pisa database}
\description{
  The package is designed to manage and display stellar tracks and
  isochrones from Pisa low-mass database.
  Data are archived at CDS and are available
  for download 
  (\url{http://cdsarc.u-strasbg.fr/viz-bin/qcat?J/A+A/540/A26}).
  
  The package includes tools to gather data from an active Internet
  connection, print and plot them with appropriate format.

  Tools for make new isocrones at desired ages, and to interpolate the
  database for composition not already available are also provided.
  
}
\references{M. Dell'Omodarme, G. Valle, S. Degl'Innocenti, and P.G. Prada Moroni
(2012). \emph{The Pisa Stellar Evolution Data Base for low-mass
  stars}. Astronomy and Astrophysics, 540, A26.}

\keyword{package}
