% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.r
\name{trim}
\alias{trim}
\title{This is a function remove space characters from the head and tail of a string}
\usage{
trim(j)
}
\description{
This is a function remove space characters from the head and tail of a string
}
\examples{
trim(' cats haunt me ')
}
\keyword{string}
