% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{stars}
\alias{stars}
\title{stars}
\usage{
stars(phrase = "code", user = NULL, language = NULL, ghname = NULL,
  newest_first = TRUE, interactive = TRUE)
}
\arguments{
\item{phrase}{Text string to search for. Defaults to "code"}

\item{user}{Name of GitHub user whose stars you want to search. Defaults to
your own profile}

\item{language}{Filter results to specified primary repository language}

\item{ghname}{Filter results to repos belonging to a specified GitHub name}

\item{newest_first}{Sort by newest stars first}

\item{interactive}{If \code{FALSE}, just print results to screen, invisibly
issue a return object, and exit.}
}
\value{
Interactive screen dump of results enabling you to select the best
match and open the corresponding GitHub repository
}
\description{
Text search GitHub stars
}
