% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{stars}
\alias{stars}
\title{stars}
\usage{
stars(phrase = "", user = whoami::gh_username(), language = NULL,
  ghname = NULL, newest_first = TRUE, interactive = TRUE)
}
\arguments{
\item{phrase}{Text string to search for}

\item{user}{Name of github user whose stars you want to search. Defaults to
your own profile (detected via \link{whoami})}

\item{language}{Filter results to specified primary repository language}

\item{ghname}{Filter results to repos belonging to a specified github name}

\item{newest_first}{Sort by newest stars first}

\item{interactive}{If \code{FALSE}, just print results to screen, invisibly
issue a return object, and exit.}
}
\value{
Interactive screen dump of results enabling you to select the best
match and open the corresponding github repository
}
\description{
Text search github stars
}
